refactor: extract canonical hidden set logic into shared utility
Moved the duplicated canonical dedup logic from inboxBadge.ts and
inboxTreeProvider.ts into a shared canonicalDedup.ts module. Both
consumers now use buildCanonicalHiddenSet() from the shared module.
